requirements.txt
Flask>=2.3.0
Pillow>=10.1.0
requests>=2.31.0
Werkzeug>=2.3.0
python-dotenv==1.0.0
openai>=1.0.0
opencv-python>=4.8.0 